PdfConvertExcel/requirements.txt
aspose-pdf~=23.9.0
PyPDF2~=3.0.1
pandas~=2.1.1